requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pydantic[email]==2.5.2
rapidfuzz==3.6.1

# Performance
httptools==0.6.0
uvloop==0.19.0; sys_platform != "win32"

# Testing
pytest==7.4.3
pytest-asyncio==0.23.0
httpx==0.25.2
